+++ spec/javascripts/serialize.spec.js
@@ -1,4 +1,26 @@
 describe('serializing a form', function() {
+
+  describe('when no elements are found', function() {
+    var View = Backbone.View.extend({
+      render: function() {
+        this.$el.html('<span>When I am alone I pretend to be a carrot<span/>');
+      }
+    });
+
+    var view;
+
+    beforeEach(function() {
+      view = new View();
+      view.render();
+    });
+
+    it('should not throw an exception', function() {
+      expect(function() {
+        Backbone.Syphon.serialize(view);
+      }).to.not.throw(Error);
+    });
+  });
+
   describe('when serializing a text input', function() {
     beforeEach(function() {
       this.View = Backbone.View.extend({
@@ -136,6 +158,20 @@ describe('serializing a form', function() {
         expect(this.result.chk).to.be.false;
       });
     });
+
+    describe('and the checkbox is indeterminate', function() {
+      beforeEach(function() {
+        this.view = new this.View();
+        this.view.render();
+        this.view.$('#the-checkbox').prop('indeterminate', true);
+
+        this.result = Backbone.Syphon.serialize(this.view);
+      });
+
+      it('should return an object with a value of null', function() {
+        expect(this.result.chk).to.be.null;
+      });
+    });
   });
 
   describe('when serializing a button', function() {
@@ -268,5 +304,90 @@ describe('serializing a form', function() {
     it('retrieves the inputs\' values', function() {
       expect(this.result.foo).to.equal('bar');
     });
+  });
+
+  describe('when given more than 1 form', function() {
+    beforeEach(function() {
+      this.View = Backbone.View.extend({
+        tagName: 'form',
+        render: function() {
+          this.$el.html(
+              '<form>' +
+              '<input type="text" name="foo" value="bar">' +
+              '</form>' +
+              '<form>' +
+              '<input type="text" name="bar" value="foo">' +
+              '</form>'
+          );
+        }
+      });
+
+      this.view = new this.View();
+      this.view.render();
+
+      this.result = Backbone.Syphon.serialize(this.view);
+    });
+
+    it('retrieves all values', function() {
+      expect(_.keys(this.result).length).to.equal(2);
+      expect(this.result.foo).to.equal('bar');
+      expect(this.result.bar).to.equal('foo');
+    });
+  });
+
+  describe('when a form tag does not exist', function() {
+    beforeEach(function() {
+      this.View = Backbone.View.extend({
+        tagName: 'form',
+        render: function() {
+          this.$el.html(
+              '<input type="text" name="foo" value="bar">' +
+              '<input type="text" name="bar" value="foo">'
+          );
+        }
+      });
+
+      this.view = new this.View();
+      this.view.render();
+
+      this.result = Backbone.Syphon.serialize(this.view);
+    });
+
+    it('retrieves all values', function() {
+      expect(_.keys(this.result).length).to.equal(2);
+      expect(this.result.foo).to.equal('bar');
+      expect(this.result.bar).to.equal('foo');
+    });
+  });
+
+  describe('when ignoring a field by selector', function() {
+    beforeEach(function() {
+      this.View = Backbone.View.extend({
+        render: function() {
+          this.$el.html(
+              '<form>' +
+              '<input name="a">' +
+              '<input name="b" class="doNotSerializeMe">' +
+              '<input name="c">' +
+              '<input name="d" class="doNotSerializeMe">' +
+              '<button name="e">' +
+              '</form>'
+          );
+        }
+      });
+
+      this.view = new this.View();
+      this.view.render();
+
+      // ignore all .doNotSerializeMe elements
+      Backbone.Syphon.ignoredTypes.push('.doNotSerializeMe');
+      this.result = Backbone.Syphon.serialize(this.view);
+    });
+
+    it('should not include fields excluded by selector', function() {
+      // note Syphon will create empty strings, not nulls.
+      expect(this.result).to.eql({a: '', c: ''});
+    });
+
   });
 });